many changes

+++ controllers/posts.js
@@ -8,11 +8,13 @@ import formatPosts from "../utils/formatPosts.js";
 const createPost = async (req, res) => {
   try {
     const { userId, description } = req.body;
+    const { file } = req;
+    if (!userId || !file || !description)
+      return res.status(400).json({ error: "Inputs required" });
     const { firstName, lastName, location, userPicturePath } =
       await User.findById(userId);
-    let postPicturePath = "";
-    if (req.file) postPicturePath = await cloudinaryUpload(req.file.path);
-    else return;
+    const { pictureUrl: postPicturePath, picturePublicId: postPictureId } =
+      await cloudinaryUpload(file.path);
     fs.unlinkSync(req.file.path);
     const newPost = new Post({
       userId,
@@ -22,14 +24,15 @@ const createPost = async (req, res) => {
       userPicturePath,
       description,
       postPicturePath,
+      postPictureId,
       likes: {},
       comments: [],
     });
-
     await newPost.save();
-    const posts = await Post.find().sort({ createdAt: -1 });
-    const formattedPosts = formatPosts(posts);
-    res.status(201).json(formattedPosts);
+
+    let post = await Post.find().sort({ createdAt: -1 }).limit(1);
+    const formattedPost = formatPosts(post);
+    res.status(201).json(formattedPost[0]);
   } catch (error) {
     console.log(error.message);
     res.status(409).json({ error: error.message });
@@ -37,7 +40,7 @@ const createPost = async (req, res) => {
 };
 
 /* READ */
-const getFeedPosts = async (req, res) => {
+const getFeedPosts = async (_, res) => {
   try {
     const posts = await Post.find().sort({ createdAt: -1 });
     const formattedPosts = formatPosts(posts);
@@ -50,7 +53,7 @@ const getFeedPosts = async (req, res) => {
 const getUserPosts = async (req, res) => {
   try {
     const { userId } = req.params;
-    const posts = await Post.find({ userId });
+    const posts = await Post.find({ userId }).sort({ createdAt: -1 });
     const formattedPosts = formatPosts(posts);
     res.status(200).json(formattedPosts);
   } catch (error) {
@@ -61,15 +64,15 @@ const getUserPosts = async (req, res) => {
 /* UPDATE */
 const likePost = async (req, res) => {
   try {
-    const { id } = req.params;
+    const { postId } = req.params;
     const { userId } = req.body;
-    const post = await Post.findById(id);
+    const post = await Post.findById(postId);
     const isLiked = post.likes.get(userId);
     if (isLiked) post.likes.delete(userId);
     else post.likes.set(userId, true);
 
     let updatedPost = await Post.findByIdAndUpdate(
-      id,
+      postId,
       { likes: post.likes },
       { new: true }
     );
@@ -82,12 +85,80 @@ const likePost = async (req, res) => {
   }
 };
 
-/* DELETE */
-const deletePost = async (req, res) => {
-  const { file } = req.body;
-  const result = await cloudinaryDelete(file);
-  if (result[result] === "ok") res.status(200).send("Deleted");
-  else res.status(404).send("File doesn't exist");
+const editPost = async (req, res) => {
+  try {
+    const { postId } = req.params;
+    const { userId, description } = req.body;
+    const result = await Post.findById(postId);
+    if (!result) return res.status(404).json({ error: "Post not found." });
+    if (!(result.userId === userId)) {
+      return res
+        .status(403)
+        .json({ error: "You are not authorized for this action!" });
+    }
+    let updatedPost = await Post.findByIdAndUpdate(
+      postId,
+      { description: description },
+      { new: true }
+    );
+
+    updatedPost = updatedPost.toJSON();
+    const formattedPost = { postId: updatedPost._id, ...updatedPost };
+    res.status(200).json(formattedPost);
+  } catch (error) {
+    res.status(500).json({ error: error.message });
+  }
 };
 
-export { createPost, getFeedPosts, getUserPosts, likePost, deletePost };
+const newComment = async (req, res) => {
+  try {
+    const { postId } = req.params;
+    const { comment } = req.body;
+    const result = await Post.findById(postId);
+    if (!result) return res.status(404).json({ error: "Post not found." });
+    const newComments = result.comments.slice();
+    newComments.push(comment);
+    console.log(newComments);
+    let updatedPost = await Post.findByIdAndUpdate(
+      postId,
+      { comments: newComments },
+      { new: true }
+    );
+
+    updatedPost = updatedPost.toJSON();
+    const formattedPost = { postId: updatedPost._id, ...updatedPost };
+    res.status(200).json(formattedPost);
+  } catch (error) {
+    res.status(500).json({ error: error.message });
+  }
+};
+
+/* DELETE */
+const deletePost = async (req, res) => {
+  try {
+    const { postId } = req.params;
+    const { userId } = req.body;
+    const result = await Post.findById(postId);
+    if (!result) return res.status(404).json({ error: "Post not found." });
+    if (result.userId === userId) {
+      const { postPictureId } = result;
+      await Promise.all([result.deleteOne(), cloudinaryDelete(postPictureId)]);
+      return res.status(200).json({ success: "Post deleted.", result });
+    } else
+      return res
+        .status(403)
+        .json({ error: "You are not authorized for this action!" });
+  } catch (error) {
+    res.status(500).json({ error: error.message });
+  }
+};
+
+export {
+  createPost,
+  getFeedPosts,
+  getUserPosts,
+  likePost,
+  newComment,
+  editPost,
+  deletePost,
+};
